DFUtils

DFUtils.complete

DFUtils.dictstodf

DFUtils.fixnothing!

DFUtils.readtypes

DFUtils.realtype

DFUtils.toReal

DFUtils.complete — Method

complete(df, cols...; replace_missing = missing)

Creates a DataFrame that has rows all combinations of cols.... Optionally fill missing values in other columns with replace_missing.

DFUtils.dictstodf — Method

dictstodf(dicts::Vector{Dict})

Create a DataFrame from a list of Dicts.

DFUtils.fixnothing! — Method

fixnothing!(df::DataFrame, col::Symbol)

Replace all nothing with missing in column and remove Nothing type.

DFUtils.readtypes — Function

readtypes(U::Union, types = DataType[])

Get a vector of the types of a Union type

DFUtils.realtype — Method

realtype(v; replace_string = true)

Finds a subtype of Real that v can be converted to. Automatically handles missings in vectors and optionally treats Strings that cannot be parsed to any Real with missing

DFUtils.toReal — Method

toReal(v; replace_string = true, threads = length(v) > 500)

Converts v to a Vector{T} where T <: Real. Automatically adds Union{T, Missing} if necessary and optionally replaces Stringss that cannot be parsed with missing.